+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - added fix for loader still spinning when atc modal is shown
 - added more preconnect and prefetch lookups to base.html
 - fixed duplicate options when the filter option remains checked when the option drawer is closed
+- collections PDP complete
 
 ## [1.20.33]
 - Removed Bolt and added Custom checkout button